fix(wasm): std and asset_cache build

+++ crates/network/src/client.rs
@@ -418,13 +418,7 @@ pub struct GameClientNetworkStats {
 
 impl Display for GameClientNetworkStats {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "{:?}ms, {}/s out, {}/s in",
-            self.latency_ms,
-            to_byte_unit(self.bytes_sent as usize),
-            to_byte_unit(self.bytes_received as usize)
-        )
+        write!(f, "{:?}ms, {}/s out, {}/s in", self.latency_ms, to_byte_unit(self.bytes_sent), to_byte_unit(self.bytes_received))
     }
 }
